+++ packages/tswebextension/src/lib/mv2/background/services/params-service.ts
@@ -67,6 +67,7 @@ export class ParamsService {
             const modifiedUrl = modifier.removeParameters(url);
 
             if (url !== modifiedUrl) {
+                context.isRemoveparamRedirect = true;
                 this.filteringLog.publishEvent({
                     type: FilteringEventType.RemoveParam,
                     data: {