+++ packages/tswebextension/test/lib/mv2/background/tabs/tabs-context.test.ts
@@ -50,37 +50,45 @@ describe('TabContext', () => {
 
     describe('updateTabInfo method', () => {
         it('should update tab info with the correct properties', () => {
-            const newTabInfo = {
+            const changeInfo = {
                 url: 'https://another.com',
                 status: 'loading',
+            };
+            const newTabInfo = {
+                ...tabInfo,
+                url: 'https://another.com',
+                title: 'Page Title',
             } as TabInfo;
 
-            const expectedInfo = { ...tabInfo, ...newTabInfo };
+            tabContext.updateTabInfo(changeInfo, newTabInfo);
 
-            tabContext.updateTabInfo(newTabInfo);
-
-            expect(tabContext.info).toEqual(expectedInfo);
-            expect(tabContext.info).toBe(tabInfo);
+            expect(tabContext.info).toEqual(newTabInfo);
+            expect(tabContext.info).toBe(newTabInfo);
+            expect(tabContext.info.title).toBe(newTabInfo.title);
             expect(tabContext.isSyntheticTab).toBe(false);
         });
 
         it('should handle cached document page initialization on tab update', () => {
-            const newTabInfo = {
+            const changeInfo = {
                 url: 'https://another.com',
                 status: 'loading',
+            };
+            const newTabInfo = {
+                ...tabInfo,
+                url: 'https://another.com',
             } as TabInfo;
 
-            tabContext.updateTabInfo(newTabInfo);
+            tabContext.updateTabInfo(changeInfo, newTabInfo);
 
             expect(tabContext.isDocumentRequestCached).toBe(true);
-            expect(documentApi.matchFrame).toBeCalledWith(newTabInfo.url);
+            expect(documentApi.matchFrame).toBeCalledWith(changeInfo.url);
         });
 
         it('should handle cached document page reload on tab update', () => {
-            const newTabInfo = { status: 'loading' } as TabInfo;
+            const changeInfo = { status: 'loading' };
 
             tabContext.isDocumentRequestCached = true;
-            tabContext.updateTabInfo(newTabInfo);
+            tabContext.updateTabInfo(changeInfo, tabInfo);
 
             expect(documentApi.matchFrame).toBeCalledWith(tabInfo.url);
         });